# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D Group AI Assistant",
    "description": "D Group AI customer-support and sales assistant: grounded RAG answers with sources, multi-turn chat sessions, recommendations, lead capture, transcripts, email and analytics. See docs/API.md for the full reference.",
    "version": "3.0.0"
  },
  "paths": {
...
    "/sessions/{session_id}/contact": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Submit the contact form from a chat",
        "description": "Contact/support form submitted from a chat. Confirms, then asks if anything else is needed.",
        "operationId": "submit_session_contact_sessions__session_id__contact_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionContactResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
    "/contact": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Standalone contact form",
        "description": "Standalone contact form (no chat session).",
        "operationId": "submit_contact_contact_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContactResponse": {
        "properties": {
          "lead_id": {
            "type": "string",
            "title": "Lead Id"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "lead_id",
          "message"
        ],
        "title": "ContactResponse"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
...
      "LeadInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "company": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 5,
            "title": "Message"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "the /ask request that led here"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "message"
        ],
        "title": "LeadInput",
        "description": "Contact form fields (validated before anything is stored)."
      },
      "MessageOut": {
        "properties": {
          "seq": {
            "type": "integer",
            "title": "Seq"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601, UTC"
          },
          "sources": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "seq",
          "role",
          "content",
          "kind",
          "created_at"
        ],
        "title": "MessageOut"
      },
...
      "SessionContactResponse": {
        "properties": {
          "lead_id": {
            "type": "string",
            "title": "Lead Id"
          },
          "session": {
            "$ref": "#/components/schemas/SessionState"
          },
          "message": {
            "$ref": "#/components/schemas/MessageOut"
          }
        },
        "type": "object",
        "required": [
          "lead_id",
          "session",
          "message"
        ],
        "title": "SessionContactResponse"
      },
...
      "SessionState": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "awaiting_idle_response",
              "closed"
            ],
            "title": "Status"
          },
          "flow": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Flow"
          },
          "close_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Close Reason"
          },
          "attempts_remaining": {
            "type": "integer",
            "title": "Attempts Remaining"
          },
          "irrelevant_limit": {
            "type": "integer",
            "title": "Irrelevant Limit"
          },
          "sound_enabled": {
            "type": "boolean",
            "title": "Sound Enabled"
          },
          "idle_check_after_seconds": {
            "type": "integer",
            "title": "Idle Check After Seconds"
          },
          "idle_close_after_seconds": {
            "type": "integer",
            "title": "Idle Close After Seconds"
          },
          "message_count": {
            "type": "integer",
            "title": "Message Count"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "status",
          "attempts_remaining",
          "irrelevant_limit",
          "sound_enabled",
          "idle_check_after_seconds",
          "idle_close_after_seconds",
          "message_count"
        ],
        "title": "SessionState"
      },
...
    }
  },
...
}